Initialize bytesused also on output VIDIOC_QBUF.
When an application submits an output buffer to the streaming interface, the
value of the bytesused field gets ignored and lost to subsequent consumers of
this buffer. This is OK for fixed-size data types, but it causes problems with
compressed sources, since they rely on this field to signal the actual size of
the data.

Thusly, this change enables using the streaming interface to transmit compressed
frame data.