Makefile: remove depmod call in modules_install target
A call to depmod is not required - modules_install already does
this for us. Also this does not make sense when we are cross-compiling
the module, because depmod will be called against the host kernel. This call
also does not work if we are cross-compiling in a container where there is no
host kernel at all.

Signed-off-by: Todor Minchev <todor.minchev@linux.intel.com>