Allow reading buffered frames
Why:

If userspace reader is blocked for some reason, it is expected that when
reading continues, frames are not dropped unless the buffer overflows.

Currently, if reader is more than one frame behind writer, next read
skips to newest frame.

It seems that allowing read from buffer is in line with how uvc v4l2
devices work.

What:

Change skipping threshold so that skipping happens only when buffer
overflows.