confine v4l2loopback_cleanup_module
In case v4l2loopback gets embedded into a full kernel tree and is
compiled as a built-in for whatever reason, the
v4l2loopback_cleanup_module routine is reported as unused because it is
called on module unload only, which is not compiled in case of built-in.

Fix this by decorating v4l2loopback_cleanup_module with another ifdef.

Signed-off-by: Oleksandr Natalenko <oleksandr@natalenko.name>