Fix repeated VIDIOC_ENUM_FRAMEINTERVALS calls
The previous code would return an error if the VIDIOC_ENUM_FRAMEINTERVALS
ioctl was invoked more than once during the lifetime of the module.
Why this was done is beyond me and I could not figure it out from the git
history either.

Fix this by checking for the index in the argp struct first and always
return -EINVAL for values greater than zero.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>