Limit v4l2_loopback_write calls to (streaming) writers
Previously whenever someone called the write function on the loopback device
the device was put in a state where no further openers could request access
to write frames to the device. While the intention of this is correct it
was missing to also make that opener doing the write operation an writer.

When an opener releases its file descriptor this output capability is
usually restored due to the fact that there's only ever one writer and
that writer should have the opener->type set to WRITER. With the
unconditional update of the dev->ready_to_output field this invariant is
broken as a file descriptor starts with a opener->type of UNKNOWN. Thus
when performing a write operation followed by closing the file descriptor
the logic to restore the output capabilities does not trigger and the
output capability is lost.

Thus instead of aiding sloppy user-land programming and guessing what
the caller desires this patch makes the write operation more strict in
its preconditions by enforcing the file descriptor to belong to an
opener with type WRITER. Failure to adhere to this is indicated by
an explicit EINVAL error return, as write operations as a non-WRITER
should not be allowed.

Signed-off-by: Benny Baumann <BenBE@geshi.org>