Revert "Fix signed integer overflows in increments."
This reverts commit 7c4ca0d207ab1338db2aae8e4a78a04a35fe25c8.

The commit introduced a bug where mod_inc returned the incremented
value, while the places it is used in expected the previous value.
Additionally, the rest of the logic for tracking buffer positions does
not work when the read_position and write_position wrap around, since
they are directly compared.

In particular, in that model, it is impossible to distinguish the case
of all buffers full vs. all buffers empty, since both pointers will be
equal in both cases. So a more thorough refactor is necessary to fix
that...

Let's revert this for now, since it breaks everything as it stands.

Fixes: #535
Signed-off-by: Asahi Lina <lina@asahilina.net>