Change the type of read_position and write_position to 64-bit
This basically solves the same problem that 7c4ca0d207a tried to solve.

It is practically impossible for a 64-bit type to actually overflow (it
would take 4 billion years of 60 FPS frames), and this avoids having to
do a much more thorough refactor to switch to wrapping buffer indices
with the extra requirement of tracking buffer fullness separately.

I checked that all of the usages should promote/truncate properly: (s64)
(OP) (s32 or u32) should always promote to 64-bit, and all of the
references to read_position and write_position are either direct
comparisons or apply a modulo before storing into a smaller type, except
for one which this commit fixes.

Signed-off-by: Asahi Lina <lina@asahilina.net>